+++ sdk/loadtesting/playwright/samples/v1/javascript/manually-connecting-to-browsers/playwright.service.config.js
@@ -1,8 +1,8 @@
-const { getServiceConfig } = require("@azure/playwright");
+const { createAzurePlaywrightConfig } = require("@azure/playwright");
 const { defineConfig } = require('@playwright/test');
 const { DefaultAzureCredential } = require("@azure/identity");
 const config = require("./playwright.config");
 
-export default defineConfig(config, getServiceConfig(config, {
-    credential: new DefaultAzureCredential()
+export default defineConfig(config, createAzurePlaywrightConfig(config, {
+  credential: new DefaultAzureCredential()
 }));