+++ sdk/loadtesting/playwright/samples/v1/typescript/using-connect-options/playwright.service.config.ts
@@ -1,4 +1,4 @@
-import { getServiceConfig } from "@azure/playwright";
+import { createAzurePlaywrightConfig } from "@azure/playwright";
 import { defineConfig } from "@playwright/test";
 import { DefaultAzureCredential } from "@azure/identity";
 import config from "./playwright.config.js";
@@ -7,7 +7,7 @@ const credential = new DefaultAzureCredential();
 
 export default defineConfig(
   config,
-  getServiceConfig(config, {
+  createAzurePlaywrightConfig(config, {
     credential,
   }),
 );