+++ sdk/monitor/monitor-opentelemetry/test/internal/unit/shared/config.test.ts
@@ -171,22 +171,18 @@ describe("Library/Config", () => {
       assert.deepStrictEqual(config.tracesPerSecond, undefined, "Wrong tracesPerSecond");
       assert.deepStrictEqual(
         config.instrumentationOptions.azureSdk?.enabled,
-        false,
+        true,
         "Wrong azureSdk",
       );
-      assert.deepStrictEqual(
-        config.instrumentationOptions.mongoDb?.enabled,
-        false,
-        "Wrong mongoDb",
-      );
-      assert.deepStrictEqual(config.instrumentationOptions.mySql?.enabled, false, "Wrong mySql");
+      assert.deepStrictEqual(config.instrumentationOptions.mongoDb?.enabled, true, "Wrong mongoDb");
+      assert.deepStrictEqual(config.instrumentationOptions.mySql?.enabled, true, "Wrong mySql");
       assert.deepStrictEqual(
         config.instrumentationOptions.postgreSql?.enabled,
-        false,
+        true,
         "Wrong postgreSql",
       );
-      assert.deepStrictEqual(config.instrumentationOptions.redis?.enabled, false, "Wrong redis");
-      assert.deepStrictEqual(config.instrumentationOptions.redis4?.enabled, false, "Wrong redis4");
+      assert.deepStrictEqual(config.instrumentationOptions.redis?.enabled, true, "Wrong redis");
+      assert.deepStrictEqual(config.instrumentationOptions.redis4?.enabled, true, "Wrong redis4");
       assert.deepStrictEqual(
         config.azureMonitorExporterOptions?.disableOfflineStorage,
         undefined,
@@ -208,7 +204,7 @@ describe("Library/Config", () => {
         },
         samplingRatio: 0.7,
         instrumentationOptions: {
-          redis4: { enabled: true },
+          redis4: { enabled: false },
         },
       };
       env["APPLICATIONINSIGHTS_CONFIGURATION_CONTENT"] = JSON.stringify(jsonOptions);
@@ -236,26 +232,22 @@ describe("Library/Config", () => {
         "Wrong connectionString",
       );
       assert.deepStrictEqual(config.instrumentationOptions.http?.enabled, false, "Wrong http");
-      assert.deepStrictEqual(config.instrumentationOptions.redis4?.enabled, true, "Wrong redis4");
+      assert.deepStrictEqual(config.instrumentationOptions.redis4?.enabled, false, "Wrong redis4");
 
       // Default values
       assert.deepStrictEqual(
         config.instrumentationOptions.azureSdk?.enabled,
-        false,
+        true,
         "Wrong azureSdk",
       );
-      assert.deepStrictEqual(
-        config.instrumentationOptions.mongoDb?.enabled,
-        false,
-        "Wrong mongoDb",
-      );
-      assert.deepStrictEqual(config.instrumentationOptions.mySql?.enabled, false, "Wrong mySql");
+      assert.deepStrictEqual(config.instrumentationOptions.mongoDb?.enabled, true, "Wrong mongoDb");
+      assert.deepStrictEqual(config.instrumentationOptions.mySql?.enabled, true, "Wrong mySql");
       assert.deepStrictEqual(
         config.instrumentationOptions.postgreSql?.enabled,
-        false,
+        true,
         "Wrong postgreSql",
       );
-      assert.deepStrictEqual(config.instrumentationOptions.redis?.enabled, false, "Wrong redis");
+      assert.deepStrictEqual(config.instrumentationOptions.redis?.enabled, true, "Wrong redis");
     });
   });